requirements.txt
absl-py==1.4.0
alembic==1.11.1
anyio==3.7.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
autograd==1.6.2
autograd-gamma==0.5.0
-e git+https://github.com/iWiley/Wistu.Web.git@58fc46b5b5c4a1221a7333b145021199a63213cd#egg=AutoModel&subdirectory=ScientificAnalysisPlatform/AutoModel-src
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
bottle==0.12.25
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
chrome-gnome-shell==0.0.0
click==8.0.3
cmaes==0.9.1
cmake==3.26.4
colorama==0.4.4
colorlog==6.7.0
comm==0.1.4
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==3.0.0
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecos==2.0.12
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
feather-format==0.4.1
filelock==3.12.2
fonttools==4.40.0
formulaic==0.6.3
fqdn==1.5.1
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.57.0
gyp==0.1
h5py==3.9.0
httplib2==0.20.2
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
inflate64==0.3.1
interface-meta==1.3.0
ipykernel==6.25.1
ipympl==0.9.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isoduration==20.11.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.7
jupyterlab_server==2.24.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lifelines==0.27.7
lit==16.0.6
littleutils==0.2.2
llvmlite==0.40.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==8.10.0
mpl-font==1.1.0
mpld3==0.5.9
mpmath==1.3.0
multivolumefile==0.2.3
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
optuna==3.2.0
optuna-dashboard==0.10.3
osqp==0.6.3
outdated==0.2.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandas-flavor==0.6.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pingouin==0.5.3
platformdirs==3.10.0
plotly==5.18.0
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.6
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycairo==1.20.1
pycox==0.2.3
pycparser==2.21
pycryptodomex==3.18.0
pycups==2.0.1
pydicom==2.4.3
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pykwalify==1.8.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyppmd==1.0.0
pyradiomics==3.1.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-json-logger==2.0.7
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
pyzstd==0.15.9
qdldl==0.1.7.post0
rchitect==0.3.40
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rpy2==3.5.14
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.2.2
scikit-survival==0.21.0
scipy==1.11.1
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.2
SimpleITK==2.2.1
six==1.16.0
sklearn==0.0.post5
sklearn-pandas==2.2.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
ssh-import-id==5.11
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2023.7.4
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.1
torchtuples==0.2.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
vboxapi==1.0
versioneer==0.29
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
widgetsnbextension==3.6.6
wrapt==1.15.0
xarray==2023.9.0
xdg==5
xkit==0.0.0
zipp==1.0.0
